# docs/openapi.yaml
openapi: "3.0.3"
info:
  title: e-ten
  version: 1.0.0
components:
...
  parameters:
...
    work_id:
      name: work_id
      in: path
      required: true
      schema:
        type: integer
...
  schemas:
...
    WorkList:
      required: [id, name, catch_copy, icon_url]
      properties:
        work_id:
          type: integer
        name:
          type: string
        genres:
          type: array
          items:
            type: string
        technologies:
          type: array
          items:
            type: string
        catch_copy:
          type: string
        icon_url:
          type: string

    # Workの詳細用のプロパティ
    WorkDetail:
      required: [description, users]
      properties:
        description:
          type: string
        images:
          type: array
          items:
            type: string
        movie:
          type: string
        works_url:
          description: 作品のURL
          type: string
        users:
          type: array
          description: チームメンバーのidと名前と役割のオブジェクトの配列
          items:
            type: object
            properties:
              user_id:
                type: integer
              usernane:
                type: string
              role:
                type: integer
    Work:
      type: object
      description: 作品
      allOf:
        - $ref: "#/components/schemas/WorkList"
        - $ref: "#/components/schemas/WorkDetail"
paths:
...
  /works/{id}:
    get:
      summary: 作品の詳細の取得
      operationId: getWork
      tags: ["work"]
      parameters:
        - $ref: "#/components/parameters/work_id"
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  # TODO:icon_urlいらない
                  - $ref: "#/components/schemas/Work"
                  - properties:
                      is_bookmarked:
                        type: boolean
                        description: ブックマークしているかどうかのフラグ

    post:
      summary: 作品の新規作成
      operationId: createWork
      tags: ["work"]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              $ref: "#/components/schemas/Work"
      responses:
        201:
          description: OK

    put:
      summary: 作品の更新
      operationId: updateWork
      tags: ["work"]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              $ref: "#/components/schemas/Work"
      responses:
        200:
          description: OK

    delete:
      summary: 作品の削除
      operationId: deleteWork
      tags: ["work"]
      parameters:
        - $ref: "#/components/parameters/work_id"
      responses:
        200:
          description: OK
...